Exception Handling

+++ service/src/main/java/ch/uzh/ifi/seal/soprafs15/service/GameServiceImpl.java
@@ -6,6 +6,7 @@ import ch.uzh.ifi.seal.soprafs15.controller.beans.game.GameResponseBean;
 import ch.uzh.ifi.seal.soprafs15.model.User;
 import ch.uzh.ifi.seal.soprafs15.model.game.Game;
 import ch.uzh.ifi.seal.soprafs15.model.repositories.GameRepository;
+import ch.uzh.ifi.seal.soprafs15.service.exceptions.GameNotFoundException;
 import ch.uzh.ifi.seal.soprafs15.service.mapper.GameMapperService;
 import ch.uzh.ifi.seal.soprafs15.model.repositories.UserRepository;
 import org.slf4j.Logger;
@@ -60,6 +61,11 @@ public class GameServiceImpl extends GameService {
     @Override
     public GameResponseBean getGame(Long gameId) {
         Game game = gameRepository.findOne(gameId);
+
+        if(game == null) {
+            throw new GameNotFoundException(gameId, GameServiceImpl.class);
+        }
+
         return gameMapperService.toGameResponseBean(game);
     }
 
@@ -68,7 +74,11 @@ public class GameServiceImpl extends GameService {
         User owner = gameMapperService.toUser(bean);
         Game game = gameRepository.findOne(gameId);
 
-        if(game.getOwner().equals(owner.getUsername()))
+        if(game == null) {
+            throw new GameNotFoundException(gameId, GameServiceImpl.class);
+        }
+
+        if(game != null && game.getOwner().equals(owner.getUsername()))
             gameRepository.delete(gameId);
     }
 }
